+++ Runtime/Scripts/Components/UIAnimations/UIDotTweenAnimation.cs
@@ -3,7 +3,7 @@ using System.Threading;
 using Cysharp.Threading.Tasks;
 using DG.Tweening;
 
-namespace com.parkminpackages.ugui.UIAnimations
+namespace Components.UIAnimations
 {
 	public abstract class UIDotTweenShowAnimation : ShowAnimation
 	{