+++ Runtime/Scripts/Components/UIAnimations/UILMSlideHideAnimation.cs
@@ -1,33 +1,44 @@
-﻿#if DOTWEEN && UNITASK_DOTWEEN_SUPPORT
+﻿#if LITMOTION_SUPPORT
 using System;
-using DG.Tweening;
+using Enums;
+using LitMotion;
 using UnityEngine;
 
-namespace com.parkminpackages.ugui.UIAnimations
+namespace Components.UIAnimations
 {
 	[DisallowMultipleComponent]
-	public class UIDTSlideShowAnimation : UIDotTweenShowAnimation
+	public class UILMSlideHideAnimation : UILitMotionHideAnimation
 	{
-		public override Tween CreateTween() {
+		public override MotionHandle CreateMotion() {
 			Vector3 canvasSize = _rootCanvasRectTransform.rect.size;
-			switch (AppearingDirection) {
+
+			Vector3 targetPos = Target.localPosition;
+
+			switch (DisappearingDirection) {
 				case Direction.Left:
-					Target.localPosition += new Vector3(-canvasSize.x, 0, 0);
+					targetPos += new Vector3(-canvasSize.x, 0, 0);
 					break;
+
 				case Direction.Right:
-					Target.localPosition += new Vector3(canvasSize.x, 0, 0);
+					targetPos += new Vector3(canvasSize.x, 0, 0);
 					break;
+
 				case Direction.Top:
-					Target.localPosition += new Vector3(0, canvasSize.y, 0);
+					targetPos += new Vector3(0, canvasSize.y, 0);
 					break;
+
 				case Direction.Bottom:
-					Target.localPosition += new Vector3(0, -canvasSize.y, 0);
+					targetPos += new Vector3(0, -canvasSize.y, 0);
 					break;
+
 				default:
 					throw new ArgumentOutOfRangeException();
 			}
 
-			return Target.DOLocalMove(_defaultLocalPos, Duration).SetEase(Ease).SetAutoKill(true);
+			return LMotion.Create(Target.localPosition, targetPos, Duration)
+			              .WithEase(Ease)
+			              .Bind(x => Target.localPosition = x)
+			              .AddTo(Target.gameObject);
 		}
 		public override void CaptureCurrent() {
 			_defaultLocalPos = Target.localPosition;
@@ -37,8 +48,8 @@ namespace com.parkminpackages.ugui.UIAnimations
 		}
 
 		public float Duration = 0.2f;
-		public Ease Ease = Ease.Unset;
-		public Direction AppearingDirection;
+		public Ease Ease = Ease.Linear;
+		public Direction DisappearingDirection;
 
 		protected override void Awake() {
 			base.Awake();