+++ Runtime/Scripts/Components/Views/GameObjectTransitionButtonEditor.cs
@@ -4,7 +4,7 @@ using UnityEditor;
 using UnityEditor.UI;
 using UnityEngine;
 
-namespace com.parkminpackages.ugui
+namespace Components.Views
 {
 	[CustomEditor(typeof(GameObjectTransitionButton))]
 	public class GameObjectTransitionButtonEditor : ButtonEditor