+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] - 2026-07-25
+
+### Breaking Changes
+- Reorganized Runtime scripts into Components, Objects, Interfaces, and Enums folders.
+- Updated namespaces to match the new folder structure.
+
 ## [2.0.0] - 2026-07-25
 
 ### Breaking Changes